backend/requirements.txt
# Lint
flake8

# Image reading
pytesseract
Pillow

# Data parsing
pandas

# Testing dependencies
pytest
pytest-cov
